# AirlineTicketingAPI/Models/UserRegistrationDto.cs
using System.ComponentModel.DataAnnotations;

namespace AirlineTicketingAPI.Models
{
    public class UserRegistrationDto
    {
        [Required]
        [StringLength(50, ErrorMessage = "Username cannot be longer than 50 characters.")]
        public string? Username { get; set; }

        [Required]
        [EmailAddress]
        [StringLength(100, ErrorMessage = "Email cannot be longer than 100 characters.")]
        public string? Email { get; set; }

        [Required]
        [StringLength(100, MinimumLength = 6, ErrorMessage = "Password must be between 6 and 100 characters.")]
        public string? Password { get; set; }
    }

}