# AirlineTicketingAPI/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AirlineTicketingAPI",
    "version": "v1"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [
          "AirlineTicketingAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Flight/{id}": {
      "get": {
        "tags": [
          "Flight"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Flight"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Flight"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Flight"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Flight"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Flight": {
      "get": {
        "tags": [
          "Flight"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "departure",
            "type": "string"
          },
          {
            "in": "query",
            "name": "arrival",
            "type": "string"
          },
          {
            "in": "query",
            "name": "date",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "pageNumber",
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32",
            "default": 10
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Flight"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Flight"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Flight"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Users/BuyTicket": {
      "post": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TicketPurchaseDetailsDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Users/register": {
      "post": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserRegistrationDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Users/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserLoginDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Flight": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "flightNumber": {
          "type": "string"
        },
        "departure": {
          "type": "string"
        },
        "arrival": {
          "type": "string"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "availableSeats": {
          "format": "int32",
          "type": "integer"
        },
        "tickets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ticket"
          }
        }
      }
    },
    "PaymentDetailsDto": {
      "required": [
        "creditCardNumber",
        "cvv",
        "expiryDate"
      ],
      "type": "object",
      "properties": {
        "creditCardNumber": {
          "format": "credit-card",
          "minLength": 1,
          "type": "string"
        },
        "expiryDate": {
          "minLength": 1,
          "pattern": "^(0[1-9]|1[0-2])\\/?([0-9]{4}|[0-9]{2})$",
          "type": "string"
        },
        "cvv": {
          "maxLength": 3,
          "minLength": 3,
          "type": "string"
        }
      }
    },
    "Ticket": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "flightID": {
          "format": "int32",
          "type": "integer"
        },
        "userID": {
          "format": "int32",
          "type": "integer"
        },
        "bookingDate": {
          "format": "date-time",
          "type": "string"
        },
        "flight": {
          "$ref": "#/definitions/Flight"
        },
        "user": {
          "$ref": "#/definitions/Users"
        }
      }
    },
    "TicketPurchaseDetailsDto": {
      "type": "object",
      "properties": {
        "ticketPurchase": {
          "$ref": "#/definitions/TicketPurchaseDto"
        },
        "paymentDetails": {
          "$ref": "#/definitions/PaymentDetailsDto"
        }
      }
    },
    "TicketPurchaseDto": {
      "required": [
        "flightID",
        "userID"
      ],
      "type": "object",
      "properties": {
        "flightID": {
          "format": "int32",
          "type": "integer"
        },
        "userID": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "UserLoginDto": {
      "required": [
        "password",
        "username"
      ],
      "type": "object",
      "properties": {
        "username": {
          "minLength": 1,
          "type": "string"
        },
        "password": {
          "minLength": 1,
          "type": "string"
        }
      }
    },
    "UserRegistrationDto": {
      "required": [
        "email",
        "password",
        "username"
      ],
      "type": "object",
      "properties": {
        "username": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "email": {
          "format": "email",
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "password": {
          "maxLength": 100,
          "minLength": 6,
          "type": "string"
        }
      }
    },
    "Users": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "tickets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ticket"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme."
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}